added user id to all requests

+++ src/ingredients/ingredients.service.ts
@@ -12,25 +12,28 @@ export class IngredientsService {
     @Inject(PRODUCT_MODEL) private productModel: Model<Product>,
   ) {}
 
-  async create(createIngredientDto: CreateIngredientDto): Promise<Ingredient> {
-    return new this.ingredientModel(createIngredientDto).save();
+  async create(createIngredientDto: CreateIngredientDto, userId: string): Promise<Ingredient> {
+    return new this.ingredientModel({
+      ...createIngredientDto,
+      userId,
+    }).save();
   }
 
-  async findAll(): Promise<Ingredient[]> {
-    return this.ingredientModel.find().exec();
+  async findAll(userId: string): Promise<Ingredient[]> {
+    return this.ingredientModel.find({ userId }).exec();
   }
 
-  async findOne(id: string): Promise<Ingredient> {
+  async findOne(id: string, userId: string): Promise<Ingredient> {
     try {
-      return await this.ingredientModel.findById(id).orFail().exec();
+      return await this.ingredientModel.findOne({ _id: id, userId }).orFail().exec();
     } catch (exception) {
       console.log(exception);
       throw new NotFoundException('Could not find ingredient');
     }
   }
 
-  async update(id: string, updateIngredientDto: UpdateIngredientDto): Promise<Ingredient> {
-    const updatedIngredient = await this.findOne(id);
+  async update(id: string, updateIngredientDto: UpdateIngredientDto, userId: string): Promise<Ingredient> {
+    const updatedIngredient = await this.findOne(id, userId);
     try {
       Object.assign(updatedIngredient, updateIngredientDto);
       return await updatedIngredient.save();
@@ -40,14 +43,16 @@ export class IngredientsService {
     }
   }
 
-  async remove(id: string) {
-    const product = await this.productModel.findOne({ 'ingredients.ingredient': new Types.ObjectId(id) }).exec();
+  async remove(id: string, userId: string) {
+    const product = await this.productModel
+      .findOne({ 'ingredients.ingredient': new Types.ObjectId(id), userId })
+      .exec();
     if (product) {
       throw new ConflictException('CONFLICT_INGREDIENT_ASSIGNED_TO_PRODUCT');
     }
 
     try {
-      await this.ingredientModel.deleteOne({ _id: id }).orFail().exec();
+      await this.ingredientModel.deleteOne({ _id: id, userId }).orFail().exec();
     } catch (exception) {
       console.log(exception);
       throw new NotFoundException('Could not find ingredient');
