+++ src/orders/orders.service.ts
@@ -10,55 +10,57 @@ import { ProductsService } from '../products';
 export class OrdersService {
   constructor(@Inject(ORDER_MODEL) private orderModel: Model<Order>, private productsService: ProductsService) {}
 
-  async create(createOrderDto: CreateOrderDto) {
+  async create(createOrderDto: CreateOrderDto, userId: string) {
     try {
-      const order = await this.convertToOrder(createOrderDto);
-      return new this.orderModel(order).save();
+      const order = await this.convertToOrder(createOrderDto, userId);
+      return new this.orderModel({ ...order, userId }).save();
     } catch (exception) {
       console.log(exception);
       throw new InternalServerErrorException('Could not create order');
     }
   }
 
-  async findAll() {
-    return this.orderModel.find().exec();
+  async findAll(userId: string) {
+    return this.orderModel.find({ userId }).exec();
   }
 
-  async findOne(id: string) {
+  async findOne(id: string, userId: string) {
     try {
-      return await this.orderModel.findById(id).orFail().exec();
+      return await this.orderModel.findOne({ _id: id, userId }).orFail().exec();
     } catch (exception) {
       console.log(exception);
       throw new NotFoundException('Could not find order');
     }
   }
 
-  async update(id: string, updateOrderDto: UpdateOrderDto) {
-    const order = await this.findOne(id);
-    const updateOrder = await this.convertToOrder(updateOrderDto);
+  async update(id: string, updateOrderDto: UpdateOrderDto, userId: string) {
+    const order = await this.findOne(id, userId);
+    const updateOrder = await this.convertToOrder(updateOrderDto, userId);
     try {
       Object.assign(order, updateOrder);
       await order?.save();
-      return this.findOne(id);
+      return this.findOne(id, userId);
     } catch (exception) {
       console.log(exception);
       throw new InternalServerErrorException('Could not update order');
     }
   }
 
-  async remove(id: string) {
+  async remove(id: string, userId: string) {
     try {
-      await this.orderModel.deleteOne({ _id: id }).orFail().exec();
+      await this.orderModel.deleteOne({ _id: id, userId }).orFail().exec();
     } catch (exception) {
       console.log(exception);
       throw new NotFoundException('Could not delete order');
     }
   }
 
-  private async convertToOrder(orderDto: CreateOrderDto | UpdateOrderDto) {
+  private async convertToOrder(orderDto: CreateOrderDto | UpdateOrderDto, userId: string) {
     if (orderDto?.products) {
       const productsDto = orderDto.products;
-      const products = await Promise.all(productsDto.map((product) => this.productsService.findOne(product.id)));
+      const products = await Promise.all(
+        productsDto.map((product) => this.productsService.findOne(product.id, userId)),
+      );
       return {
         deliveryDate: orderDto.deliveryDate,
         products: products.map((product) => ({
